+++ src/Rules/Nullable.php
@@ -1,23 +1,23 @@
 <?php
 
 namespace Expay\Refine\Rules;
-use Expay\Refine\Exceptions\InvalidField;
 
-class NotEmpty extends Rule
+/**
+ * Nullable
+ */
+class Nullable extends Rule
 {
   /**
-   * apply: Throw an error if the field is empty
+   * apply: Remove any nasty characters from the supplied string
    *
    * @param  mixed $value
    * @param  mixed $key
    * @param  mixed $request
    * @param  mixed $validationRules
-   * @return void
+   * @return string
    */
   public function apply($value, string $key="", array $request=[], array $validationRules=[])
   {
-    if (!array_key_exists($key, $request))
-      throw new InvalidField("Field '$key' should not be empty");
     return $value;
   }
 }